+++ tests/testthat/test_see_data_funcs.R
@@ -49,3 +49,20 @@ test_that("Functions to see optimal genotype and GEBV work", {
   
   clear.simdata()
 })
+
+test_that("Functions to see optimal genotype and GEBV of a group work", {
+  capture_output(g <- load.data("helper_genotypes.txt", "helper_map.txt", "helper_eff.txt"), print=F)
+  
+  expect_identical(see.optimal.haplotype(), "TAA")
+  expect_equal(see.optimal.GEBV(), 1.8)
+  expect_equal(see.minimum.GEBV(),-2.8)
+  
+  expect_identical(see.optimal.possible.haplotype(g), "TAA")
+  expect_equal(see.optimal.possible.GEBV(g), 1.8)
+  
+  g2 <- make.group(c(4L,5L))
+  expect_identical(see.optimal.possible.haplotype(g2), "TAT")
+  expect_equal(see.optimal.possible.GEBV(g2), 1.4)
+  
+  clear.simdata()
+})